% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mfx_functions.R
\name{CalcMfx}
\alias{CalcMfx}
\title{Calculate marginal effects and ICE curves}
\usage{
CalcMfx(object, X, pred_fun = predict, predictors = colnames(X),
  max_pts = 100, min_pts = 0, dydx_mean = FALSE, ...)
}
\arguments{
\item{object}{a model object}

\item{X}{an object of class \code{data.frame} containing data from which
marginal effects should be calculated}

\item{pred_fun}{a function that accepts two arguments corresponding to
\code{object} and \code{X}, above. The function should return a vector of
predicted responses. If this argument is not passed, the generic
\code{predict} function will be tried.}

\item{predictors}{a character vector of column names of \code{X} for which
marginal effects are to be calculated}

\item{max_pts}{an integer value or \code{NA}. If an integer, then \code{max_pts}
number of values will be used to calculate marginal effects and ICE curves.
If \code{max_pts} is \code{NA}, then each predictor's unique values will be used.}

\item{min_pts}{a positive integer. Predictorswith fewer than \code{min_pts}
unique values will be evaluated only at their unique values, rather than
\code{max_pts} values. If you want to guarantee all points will be evaluated
at \code{max_pts} values, set \code{min_pts} to 0.}

\item{dydx_mean}{logical indicating whether to calculate the marginal effect
based on the mean of the derivative of ICE curves. If \code{FALSE}, then
the marginal effect is calculated at each actual observation then averaged.}

\item{...}{arguments to be passed on to \code{textmineR}'s function
code{\link[textmineR]{TmParallelApply}}.}
}
\value{
Returns a \code{list} of class \code{Mfx} or \code{Mfx_list} depending on
whether the \code{predictors} argument is of length 1 or greater. If length 1,
then the result is of class \code{Mfx}. If greater than 1, the result is of
class \code{Mfx_list}, each element of which is of class \code{Mfx}. An object
of class \code{Mfx} has the following slots:
}
\description{
calculate marginal effects of features on predicted outcomes
from an arbitrary prediction model. Also calculates the necessary components
to construct individual conditional expectation (ICE) curves.
}
\details{
details coming soon
}
\section{Slots}{

\describe{
\item{\code{mfx}}{the calculated marginal effect}

\item{\code{se}}{the standard error of the calculated marginal effect}

\item{\code{conf}}{a 95% confidence interval around the effect}

\item{\code{dy}}{a vector of changes in the predicted value}

\item{\code{dx}}{a vector of changes in the predictor (set by \code{max_pts})}

\item{\code{x}}{the predictor vector set by \code{max_pts}}

\item{\code{yh0}}{inital prediction corresponding to the first value of \code{x}}

\item{\code{true_values}}{\code{data.frame} with values for \code{dx}, \code{dy},
and predictions for the actual data points.}

\item{\code{varname}}{the name of the predictor for which a marginal effect has been
calculated}
}}

\examples{
# examples coming soon


}
