% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mfx_functions.R
\name{Factor2Binary}
\alias{Factor2Binary}
\title{easily convert a k-level factor variable to k binary variables}
\usage{
Factor2Binary(y)
}
\arguments{
\item{y}{a factor variable with at least 2 levels}
}
\value{
a matrix of k binary columns, one corresponding to each level of
the input, \code{y}
}
\description{
easily convert a k-level factor variable to k binary variables
}
\examples{
my_factor <- c("bob", "sharron", "bob", "alice", "alice")

my_factor <- as.factor(my_factor)

my_binary_matrix <- Factor2Binary(my_factor)

my_binary_matrix
}
