% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maRginal.R
\docType{package}
\name{maRginal}
\alias{maRginal}
\alias{maRginal-package}
\title{maRginal}
\description{
Marginal Effects for Arbitrary Predictive Models
}
\details{
Calculate marginal effects, similar to linear coefficients, for any
predictive model. Also calculates and plots individual conditional expectation
(ICE) curves. As of this writing, the concept is experimental. Caveat emptor!
}
