% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mfx_functions.R
\name{plot.Mfx_list}
\alias{plot.Mfx_list}
\title{plot method for objects of class \code{Mfx_list}}
\usage{
\method{plot}{Mfx_list}(mfx_list, ask = TRUE, ...)
}
\arguments{
\item{mfx_list}{an object of class \code{Mfx_list}}

\item{ask}{do you want to pause and wait for user input between plots?}

\item{...}{other parameters to be passed to \code{plot.Mfx} or
\code{\link[base]{plot}}.}
}
\description{
plot method for objects of class \code{Mfx_list}
}
\details{
\code{plot.Mfx_list} simply loops over every element of \code{mfx_list} and
callsj \code{plot.Mfx} for each element. It pauses between plots depending on
the value passed to \code{ask}.
}
\examples{
# TO DO
}
