% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mfx_functions.R
\name{print.Mfx_list}
\alias{print.Mfx_list}
\title{print method for objects of class \code{Mfx_list}}
\usage{
\method{print}{Mfx_list}(mfx_list)
}
\arguments{
\item{mfx_list}{an object of class \code{Mfx_list}}
}
\description{
print method for objects of class \code{Mfx_list}
}
\details{
simply calls \code{summary.Mfx_list} with option \code{print = TRUE}
}
\examples{
# TO DO
}
