% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mfx_functions.R
\name{summary.Mfx_list}
\alias{summary.Mfx_list}
\title{summary method for objects of class \code{Mfx_list}}
\usage{
\method{summary}{Mfx_list}(mfx_list, print = TRUE)
}
\arguments{
\item{mfx_list}{an object of class \code{Mfx_list}}

\item{print}{do you want to print the result to the console?}
}
\value{
a \code{data.frame} with 6 columns:
}
\description{
summary method for objects of class \code{Mfx_list}
}
\section{Slots}{

\describe{
\item{\code{variable}}{the name of the variable or predictor}

\item{\code{effect}}{the marginal effect of that variable on the outcome}

\item{\code{se}}{the standard error of the the effect}

\item{\code{t_stat}}{a t-statistic on the effect with a null hypothesis of
\code{effect = 0}}

\item{\code{95_conf_low}}{the lower bound of a 95% confidence interval}

\item{\code{95_conf_high}}{the upper bound of a 95% confidence interval}
}}

\examples{
# TO DO
}
